+++ src/gdalio.cpp
@@ -935,7 +935,7 @@ bool SpatRaster::from_gdalMEM(GDALDatasetH hDS, bool set_geometry, bool get_valu
 
 
 
-char ** set_GDAL_options(std::string driver, double diskNeeded, bool writeRGB, bool parallel, std::vector<std::string> gdal_options) {
+char ** set_GDAL_options(std::string driver, double diskNeeded, bool writeRGB, bool parallel, unsigned threads, std::vector<std::string> gdal_options) {
 
 	char ** gdalops = NULL;
 	if (driver == "GTiff") {
@@ -974,7 +974,11 @@ char ** set_GDAL_options(std::string driver, double diskNeeded, bool writeRGB, b
 				}
 			}
 			if (numt) {
-				gdalops = CSLSetNameValue( gdalops, "NUM_THREADS", "ALL_CPUS");				
+				if (threads > 0) {
+					gdalops = CSLSetNameValue( gdalops, "NUM_THREADS", std::to_string(threads).c_str());
+				} else {
+					gdalops = CSLSetNameValue( gdalops, "NUM_THREADS", "ALL_CPUS");
+				}
 			}
 		}
 		if (writeRGB) {
@@ -1020,7 +1024,7 @@ bool SpatRaster::create_gdalDS(GDALDatasetH &hDS, std::string filename, std::str
 			return(false);
 		}
 
-		papszOptions = set_GDAL_options(driver, diskNeeded, false, opt.parallel, opt.gdal_options);
+		papszOptions = set_GDAL_options(driver, diskNeeded, false, opt.parallel, opt.threads, opt.gdal_options);
 
 		if (datatype == "INT4S") {
 			naflag = INT32_MIN; //-2147483648;